Rename main to create
