+++ conda_lockfiles/cli.py
@@ -1,5 +1,5 @@
 """
-conda pip subcommand for CLI
+conda lockfiles subcommand for CLI
 """
 
 from __future__ import annotations
@@ -31,7 +31,7 @@ def configure_parser(parser: argparse.ArgumentParser):
 
 def execute(args: argparse.Namespace) -> int:
     from conda.base.context import context, determine_target_prefix
-    from .main import create_environment_from_lockfile
+    from .create import create_environment_from_lockfile
 
     prefix = determine_target_prefix(context, args)
     create_environment_from_lockfile(
